requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==2.2
django-ckeditor==6.3.0
django-crispy-forms==1.14.0
django-filter==21.1
django-js-asset==2.0.0
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
